Refactor: split out DisposeCall WIP.
Needs update in Gu.Roslyn.Extensions.
