+++ IDisposableAnalyzers.Test/IDISP002DisposeMemberTests/HappyPath.cs
@@ -1,12 +1,13 @@
-﻿#pragma warning disable SA1203 // Constants must appear before fields
+#pragma warning disable SA1203 // Constants must appear before fields
 namespace IDisposableAnalyzers.Test.IDISP002DisposeMemberTests
 {
     using Gu.Roslyn.Asserts;
+    using Microsoft.CodeAnalysis.Diagnostics;
     using NUnit.Framework;
 
     internal partial class HappyPath
     {
-        private static readonly FieldAndPropertyDeclarationAnalyzer Analyzer = new FieldAndPropertyDeclarationAnalyzer();
+        private static readonly DiagnosticAnalyzer Analyzer = new FieldAndPropertyDeclarationAnalyzer();
 
         private const string DisposableCode = @"
 namespace RoslynSandbox
